requirements.txt
Django==1.6.2
South==0.8.4
amqp==1.4.5
anyjson==0.3.3
apns-client==0.1.8
billiard==3.3.0.17
celery==3.1.11
cffi==0.8.2
cryptography==0.3
django-braces==1.4.0
django-celery==3.1.10
django-oauth-toolkit==0.7.0
djangorestframework==2.3.13
kombu==3.0.15
librabbitmq==1.5.0
oauthlib==0.6.1
psycopg2==2.5.2
pyOpenSSL==0.14
pycparser==2.10
python-dateutil==2.2
pytz==2014.2
six==1.6.1
wsgiref==0.1.2
